Add one-bit semaphore abstraction.
This is used to replace the mutex and condition variable for
transferring control between the child and the parent.  It currently has
two implementations:

 - umtx-based on FreeBSD
 - Futex-based on Linux

In theory, this could be implemented with POSIX semaphores but it's not
clear what the security implications would be.

This also fixes some bugs introduced with the move to `Handle`s.  In a
few places `Handle`s were being constructed for file descriptors that
were already owned by another `Handle`.  This caused the file
descriptor to be prematurely closed.  The `Handle` constructor that
takes a raw file descriptor is now `explicit`, which prevents this from
happening.