Move the C++ API code out for testing.
The C++ wrapper code will not be used for Verona, the generic code now
only contains things that Verona will need and is easier for veronac to
plug into.

The ExportedLibrary class is now entirely static (we only have one per
exported library) and the ExportedLibraryPrivate class is gone and all
of its logic moved into an anonymous namespace in the library runner.

Sandboxes now just need to provide init and dispatcher functions.  The
init function doesn't need to do anything, the dispatcher takes an int
containing the vtable index and a pointer containing the argument frame.
It is entirely up to consumers of the sandbox library how they lay
things out in the argument frame for any given vtable.  The C++ code
provides an example, Verona will want to do something different.