Initial implementation of seccomp-bpf sandbox.
This is still quite rough around the edges and is missing tests and
upcalls for most system calls and has several important FIXMEs.  This
may well have broken the FreeBSD build (to be fixed next week) and the
synchronisation code for sandbox launch makes me sad and will probably
be rewritten at some point.

The seccomp-bpf system allows userspace to provide eBPF filters that
define the system calls that can be authorised.  This is not ideal for
sandboxing for several reasons, most notably:

First the eBPF cannot read userspace pointers and so cannot make
decisions based on the path.  This is not possible to add safely with
the current structure of seccomp-bpf because the kernel does not
deep-copy arguments before the filter runs and so would be subject to
TOCTOU vulnerabilities (e.g. read the path to `open`, approve it, the
sandboxed process then modifies it in a different thread to one the
policy would not have allowed).  This composes poorly with the recent
trend in Linux towards system calls designed for future extensibility
that take more complex data structures as arguments.

Second, it can't handle dynamic properties and doesn't have any way of
attaching metadata to file descriptors that can be queried from the
eBPF, so you cannot easily restrict operations on individual file
descriptors (for example, allowing `ioctl` on a file descriptor for a
specific device node but no others).

This means that, for even basic usage, we need to provide some upcall
mechanism if we want to be able to do simple things like open a file.
The glibc run-time linker does not allow us to either provide file
descriptors or to interpose on its system calls, so we also need to be
able to trap and emulate these calls.

This implementation configures the seccomp-bpf filter to trap on certain
system calls (for example, `open`).  We then catch the signal that's
delivered, copy the arguments out of the trap frame, perform an upcall
to the parent process, and inject the result into the trap frame and
advance the program counter past the `syscall` instruction.  In this
commit, only `open` and `stat` are proxied but the infrastructure now
exists for emulating the others.  This will also be useful for
transparent execution of non-Capsicum-aware libraries on FreeBSD, though
most such software uses the libc wrappers around the system calls, which
we can trivially replace by providing implementations in the main
executable.

The upcall mechanism should support reentrancy in theory, but this is
not yet tested and so is almost certainly broken.

To support the emulation of `open`, this commit adds an initial
abstraction for an exported file tree.  This will be necessary for
Capsicum as well, though can reside inside the sandbox in the Capsicum
implementation.