Fix release build.
Various things were in `assert` statements that we want in release
builds and so trying a release build gave a lot of unused-variable
errors.  This commit adds a `sandbox::invariant` template function (and
some helper macros for using it) that allows invariants to be checked
easily in debug/release/any build configurations and to print an
informative message if the invariants are violated.